07-slo-scheduling/requirements.txt
httpx>=0.27
pyyaml>=6.0
numpy>=1.26
matplotlib>=3.9
pandas>=2.2
pytest>=8.0
pytest-asyncio>=0.24
